undo unrelated formatting
Drop all of the unused packages

drop build_runner, too

move shelf import to isolated

update import for test

another unrelated revert
