+++ packages/flutter_tools/pubspec.yaml
@@ -13,7 +13,6 @@ dependencies:
   archive: 3.6.1
   args: 2.7.0
   dds: 5.0.3
-  build_runner: ^2.5.4
   dwds: 24.4.0
   code_builder: 4.10.1
   collection: 1.19.1
@@ -46,8 +45,6 @@ dependencies:
   stream_channel: 2.1.4
   shelf_web_socket: 2.0.1
   shelf_static: 1.1.3
-  shelf_router: ^1.1.4
-  shelf_test_handler: ^2.0.2
   pub_semver: 2.2.0
   pool: 1.5.1
   path: 1.9.1
@@ -116,7 +113,6 @@ dependencies:
   web: 1.1.1
   web_socket: 1.0.1
   yaml_edit: 2.2.2
-  mockito: ^5.4.6
 
 dev_dependencies:
   file_testing: 3.0.2
